requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
pytest==8.4.0
pytest-cov==6.2.1
requests==2.32.5
selenium==4.34.2
virtualenv==20.34.0

